cyclic

+++ simulation.py
@@ -37,7 +37,7 @@ class Simulation:
         return 3 ** self.n
 
     def _initialize_states(self):
-        self.states = np.array([int(self._max_neighbor_count() ** (i)) for i in range(self.state_count)], dtype=np.uint32)
+        self.states = np.array([int(self._max_neighbor_count() ** (i)) for i in range(self.state_count)], dtype=np.uint64)
         for i in range(len(self.states)):
             self.states_dict[self.states[i]] = i
 
@@ -58,7 +58,7 @@ class Simulation:
         pad = 1
         padded = np.pad(self.grid, [(pad, pad)] * self.n, mode='constant', constant_values=0)
 
-        self.neighbors_grid = np.zeros_like(self.grid, dtype=np.uint32)
+        self.neighbors_grid = np.zeros_like(self.grid, dtype=np.uint64)
 
         for off in self.offsets:
             slices = []
